pyproject.toml
[tool.poetry]
name = "ascii-theater"
version = "0.1.0"
description = ""
authors = ["Meir <maornet@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
opencv-python = "^4.9.0.80"
pillow = "^10.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
